Split file related endpoints to individual files
